# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Healthcare Platform API
  version: 1.0.0
  description: |
    Comprehensive API for the Global Healthcare SaaS Platform providing telemedicine,
    patient management, appointments, clinical documentation, and billing services.

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain tokens via `/auth/login` or `/auth/register`.

    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per IP address.

  contact:
    name: API Support
    email: support@unifiedhealthcare.com
  license:
    name: Proprietary
    url: https://unifiedhealthcare.com/terms
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api-dev.unifiedhealthcare.com/api/v1
    description: Development environment
  - url: https://api-staging.unifiedhealthcare.com/api/v1
    description: Staging environment
  - url: https://api.unifiedhealthcare.com/api/v1
    description: Production environment
...
security:
  - BearerAuth: []
paths:
...
  /patients:
    post:
      tags:
        - Patients
      summary: Create patient record
      description: Create a new patient medical record
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
      responses:
        '201':
          description: Patient created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /patients/{id}:
    get:
      tags:
        - Patients
      summary: Get patient by ID
      description: Retrieve patient medical record
      operationId: getPatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Patients
      summary: Update patient record
      description: Update patient medical information
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
      responses:
        '200':
          description: Patient updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /subscriptions:
    post:
      tags:
        - Subscriptions
      summary: Create subscription
      description: Create a new subscription
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - planId
                - paymentMethodId
              properties:
                planId:
                  type: string
                  format: uuid
                paymentMethodId:
                  type: string
                couponCode:
                  type: string
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /subscriptions/{id}:
    delete:
      tags:
        - Subscriptions
      summary: Cancel subscription
      description: Cancel an active subscription
      operationId: cancelSubscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Subscription cancelled successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==========================================
  # Billing Endpoints
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
...
    CreatePatientRequest:
      type: object
      required:
        - userId
        - dateOfBirth
        - gender
      properties:
        userId:
          type: string
          format: uuid
        dateOfBirth:
          type: string
          format: date
          example: "1990-01-01"
        gender:
          type: string
          enum: [male, female, other, prefer-not-to-say]
        bloodType:
          type: string
          example: "O+"
        allergies:
          type: array
          items:
            type: string
          example: ["Penicillin", "Peanuts"]
        emergencyContact:
          type: object
          properties:
            name:
              type: string
            relationship:
              type: string
            phone:
              type: string
    UpdatePatientRequest:
      type: object
      properties:
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other, prefer-not-to-say]
        bloodType:
          type: string
        allergies:
          type: array
          items:
            type: string
        emergencyContact:
          type: object
          properties:
            name:
              type: string
            relationship:
              type: string
            phone:
              type: string
    PatientResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        medicalRecordNumber:
          type: string
          example: "MRN-123456"
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        bloodType:
          type: string
        allergies:
          type: array
          items:
            type: string
        emergencyContact:
          type: object
          properties:
            name:
              type: string
            relationship:
              type: string
            phone:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ==========================================
    # Appointment Schemas
    # ==========================================
...
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, cancelled, expired, past_due]
        currentPeriodStart:
          type: string
          format: date-time
        currentPeriodEnd:
          type: string
          format: date-time
        cancelAtPeriodEnd:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ==========================================
    # Audit & Consent Schemas
    # ==========================================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time

  # ==========================================
  # Reusable Responses
  # ==========================================
  responses:
    BadRequestError:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Bad Request"
            message: "Invalid input data"
            timestamp: "2024-01-01T00:00:00Z"
    UnauthorizedError:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Authentication token is missing or invalid"
            timestamp: "2024-01-01T00:00:00Z"
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Forbidden"
            message: "You do not have permission to access this resource"
            timestamp: "2024-01-01T00:00:00Z"
    NotFoundError:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not Found"
            message: "The requested resource was not found"
            timestamp: "2024-01-01T00:00:00Z"
...